ref: move legacy files to the new folder structure
