+++ src/app/features/auth/pages/index.ts
@@ -1,1 +1,2 @@
-export {LoginPageComponent} from './login-page/login-page.component';
+export { LoginPageComponent } from './login-page/login-page.component';
+export { RegisterPageComponent } from './register-page/register-page.component';
