+++ src/app/features/auth/pages/register-page/register-page.component.html
@@ -1,0 +1,41 @@
+<auth-background>
+  <p-card>
+    <div class="container">
+      <img alt="Logo of Inventio" class="logo" src="/assets/logo.svg">
+      <h2>Zarejestruj się</h2>
+      <form (submit)="handleSubmit()" [formGroup]="form" class="form">
+        <span>
+          <ui-label for="login" label="Login">
+            <input class="input" formControlName="login" id="login" pInputText type="text">
+          </ui-label>
+          @if (form.get('login')?.invalid && form.get('login')?.touched) {
+            <small class="error">Login nie może być pusty.</small>
+          }
+        </span>
+        <span>
+          <ui-label for="password" label="Hasło">
+            <p-password class="input" feedback="false" fluid formControlName="password" id="password" toggleMask/>
+          </ui-label>
+          @if (form.get('password')?.invalid && form.get('password')?.touched) {
+            <small class="error">Hasło nie może być puste.</small>
+          }
+        </span>
+        <span>
+          <ui-label for="passwordConfirmation" label="Powtórz hasło">
+            <p-password class="input" feedback="false" fluid formControlName="passwordConfirmation"
+                        id="passwordConfirmation" toggleMask/>
+          </ui-label>
+          @if (form.get('passwordConfirmation')?.invalid && form.get('passwordConfirmation')?.touched) {
+            <small class="error">Powtórz hasło.</small>
+          }
+          @if (form.hasError('isPasswordMatching') && form.get('passwordConfirmation')?.touched) {
+            <small class="error">Hasła się nie zgadzają.</small>
+          }
+
+        </span>
+        <p-button fluid type="submit">Zarejestruj się</p-button>
+      </form>
+      <a href="/sign-in" rel="stylesheet">Zaloguj się</a>
+    </div>
+  </p-card>
+</auth-background>
