+++ src/app/features/auth/pages/register-page/register-page.component.scss
